build: make libiconv and libintl optional
While most packages depend on libiconv and/or libintl, the are some
which do not, such as winpthreads and ncurses.

If user wants to build just such packages, they would end up building
both libiconv and libintl twice, which is an overkill and time consuming.

This commit makes both libiconv and libintl optional, and introduces
WITH_LIBICONV and WITH_LIBINTL variables to control whether they are built.

Document which packages depend on libiconv and libintl in PACKAGES.md.